Refactored rest of main components, added hooks for them to meet size preference
