+++ client/src/components/HomeDashboard.tsx
@@ -1,11 +1,13 @@
 "use client";
-import { useEffect, useState, useCallback } from "react";
-import { motion, AnimatePresence } from "framer-motion";
-import { getSupabase } from "@/lib/supabase";
+import { useEffect, useState } from "react";
 import { useStats } from "@/lib/useStats";
 import { WORLDS, type WorldId } from "@/lib/avatarData";
 import { formatDuration } from "@/lib/format";
-import type { Profile, Task } from "@/lib/types";
+import { useTasks } from "@/hooks/useTasks";
+import { useOnlineFriends } from "@/hooks/useOnlineFriends";
+import TaskSection from "./TaskSection";
+import FriendsOnlineSection from "./FriendsOnlineSection";
+import type { Profile } from "@/lib/types";
 import type { Socket } from "socket.io-client";
 
 interface Props {
@@ -61,16 +63,22 @@ export default function HomeDashboard({
   onOpenStats,
 }: Props) {
   const [selectedWorld, setSelectedWorld] = useState<WorldId>("forest");
-  const [tasks, setTasks] = useState<Task[]>([]);
-  const [newTask, setNewTask] = useState("");
   const [profileMenuOpen, setProfileMenuOpen] = useState(false);
-  const sb = getSupabase();
   const { personalStats, loading, fetchStats } = useStats(profile.id);
 
-  const [friends, setFriends] = useState<Profile[]>([]);
-  const [onlineFriendIds, setOnlineFriendIds] = useState<Set<string>>(
-    new Set(),
-  );
+  const {
+    tasks,
+    newTask,
+    setNewTask,
+    addTask,
+    toggleTask,
+    deleteTask,
+    pendingTasks,
+    completedTasks,
+    clearCompleted,
+  } = useTasks(profile.id);
+
+  const { friends, onlineFriendIds } = useOnlineFriends(profile.id, socketRef);
 
   const displayName = profile.display_name ?? profile.username ?? "You";
   const initial = displayName.charAt(0).toUpperCase();
@@ -80,118 +88,9 @@ export default function HomeDashboard({
     fetchStats();
   }, [fetchStats]);
 
-  // Fetch accepted friends list
-  const fetchFriends = useCallback(async () => {
-    const { data } = await sb
-      .from("friendships")
-      .select(
-        `
-        id, requester_id, addressee_id,
-        requester:profiles!friendships_requester_id_fkey(id, username, display_name, current_session_id, current_world_id),
-        addressee:profiles!friendships_addressee_id_fkey(id, username, display_name, current_session_id, current_world_id)
-      `,
-      )
-      .eq("status", "accepted");
-    if (data) {
-      const myId = profile.id;
-      setFriends(
-        data.map((f: any) =>
-          f.requester_id === myId ? f.addressee : f.requester,
-        ) as Profile[],
-      );
-    }
-  }, [sb, profile.id]);
-
-  useEffect(() => {
-    fetchFriends();
-  }, [fetchFriends]);
-
-  // Query online status via socket + listen for presence updates
-  useEffect(() => {
-    const socket = socketRef.current;
-    if (!socket || friends.length === 0) return;
-
-    const friendIds = friends.map((f) => f.id);
-    socket.emit("get_online_friends", { friendIds }, (online: string[]) => {
-      setOnlineFriendIds(new Set(online));
-    });
-
-    const handlePresence = ({
-      userId,
-      online,
-    }: {
-      userId: string;
-      online: boolean;
-    }) => {
-      setOnlineFriendIds((prev) => {
-        const next = new Set(prev);
-        if (online) next.add(userId);
-        else next.delete(userId);
-        return next;
-      });
-    };
-
-    socket.on("presence_update", handlePresence);
-    return () => {
-      socket.off("presence_update", handlePresence);
-    };
-  }, [socketRef, friends]);
-
-  const WORLD_LABEL: Record<string, { emoji: string; label: string }> =
-    Object.fromEntries(
-      WORLDS.map((w) => [w.id, { emoji: w.emoji, label: w.label }]),
-    );
-
   const onlineFriends = friends.filter(
     (f) => onlineFriendIds.has(f.id) || !!f.current_session_id,
   );
-
-  const fetchTasks = useCallback(async () => {
-    const { data } = await sb
-      .from("tasks")
-      .select("*")
-      .eq("owner_id", profile.id)
-      .is("room_code", null)
-      .order("created_at", { ascending: true });
-    if (data) setTasks(data as Task[]);
-  }, [sb, profile.id]);
-
-  useEffect(() => {
-    fetchTasks();
-  }, [fetchTasks]);
-
-  const addTask = async () => {
-    const text = newTask.trim();
-    if (!text) return;
-    const { data } = await sb
-      .from("tasks")
-      .insert({ owner_id: profile.id, content: text })
-      .select()
-      .single();
-    if (data) setTasks((p) => [...p, data as Task]);
-    setNewTask("");
-  };
-
-  const toggleTask = async (id: string, done: boolean) => {
-    await sb.from("tasks").update({ is_done: done }).eq("id", id);
-    setTasks((p) => p.map((t) => (t.id === id ? { ...t, is_done: done } : t)));
-  };
-
-  const deleteTask = async (id: string) => {
-    await sb.from("tasks").delete().eq("id", id);
-    setTasks((p) => p.filter((t) => t.id !== id));
-  };
-
-  const pendingTasks = tasks.filter((t) => !t.is_done);
-  const completedTasks = tasks.filter((t) => t.is_done);
-
-  const clearCompleted = async () => {
-    const done = completedTasks;
-    for (const t of done) {
-      await sb.from("tasks").delete().eq("id", t.id);
-    }
-    setTasks((p) => p.filter((t) => !t.is_done));
-  };
 
   const greetingHour = new Date().getHours();
   const greeting =
@@ -206,9 +105,8 @@ export default function HomeDashboard({
       className="min-h-screen bg-gray-900 flex flex-col"
       onClick={() => setProfileMenuOpen(false)}
     >
-      {/* Top bar — same grid as game screen, Duodoro always dead center */}
+      {/* Top bar */}
       <div className="grid grid-cols-[1fr_auto_1fr] items-center px-4 py-2.5 bg-gray-800/80 border-b border-gray-700">
-        {/* Left: Friends — right-aligned toward Duodoro */}
         <div className="flex items-center justify-end pr-2">
           <button
             onClick={(e) => {
@@ -221,7 +119,6 @@ export default function HomeDashboard({
           </button>
         </div>
 
-        {/* Center: Duodoro + status dot — always dead center */}
         <div className="flex flex-col items-center px-3 py-0.5">
           <span className="text-white font-black font-mono tracking-widest text-sm">
             Duodoro
@@ -236,7 +133,6 @@ export default function HomeDashboard({
           />
         </div>
 
-        {/* Right: Stats, Account — left-aligned from Duodoro, Account pushed far right */}
         <div className="flex items-center gap-1.5 pl-2">
           <button
             onClick={(e) => {
@@ -304,7 +200,6 @@ export default function HomeDashboard({
       {/* Main content */}
       <div className="flex-1 overflow-y-auto">
         <div className="max-w-lg mx-auto px-4 py-6 space-y-6">
-          {/* Greeting */}
           <div>
             <h1 className="text-2xl font-bold text-white font-mono">
               {greeting}, {displayName}
@@ -314,7 +209,6 @@ export default function HomeDashboard({
             </p>
           </div>
 
-          {/* Quick Stats */}
           {!loading && personalStats && (
             <div className="flex gap-2">
               <QuickStat
@@ -342,185 +236,25 @@ export default function HomeDashboard({
             </div>
           )}
 
-          {/* Goals / Tasks */}
-          <div className="bg-gray-800/50 rounded-2xl border border-gray-700 p-4">
-            <div className="flex items-center justify-between mb-3">
-              <h2 className="text-sm font-bold font-mono text-gray-400 uppercase tracking-wider">
-                Goals
-              </h2>
-              {tasks.length > 0 && (
-                <div className="flex items-center gap-2">
-                  <span className="text-[10px] font-mono text-gray-600">
-                    {completedTasks.length}/{tasks.length} done
-                  </span>
-                  {completedTasks.length > 0 && (
-                    <button
-                      onClick={clearCompleted}
-                      className="text-[10px] font-mono text-gray-600 hover:text-red-400 transition-colors"
-                    >
-                      Clean
-                    </button>
-                  )}
-                </div>
-              )}
-            </div>
+          <TaskSection
+            tasks={tasks}
+            pendingTasks={pendingTasks}
+            completedTasks={completedTasks}
+            newTask={newTask}
+            setNewTask={setNewTask}
+            addTask={addTask}
+            toggleTask={toggleTask}
+            deleteTask={deleteTask}
+            clearCompleted={clearCompleted}
+          />
 
-            {/* Task list */}
-            <div className="space-y-1.5 max-h-48 overflow-y-auto">
-              <AnimatePresence mode="popLayout">
-                {pendingTasks.map((task) => (
-                  <motion.div
-                    key={task.id}
-                    layout
-                    initial={{ opacity: 0, y: 8 }}
-                    animate={{ opacity: 1, y: 0 }}
-                    exit={{ opacity: 0, x: -16 }}
-                    className="flex items-center gap-2.5 group"
-                  >
-                    <button
-                      onClick={() => toggleTask(task.id, true)}
-                      className="flex-shrink-0 w-4.5 h-4.5 w-[18px] h-[18px] rounded border-2 border-gray-600 hover:border-emerald-500 flex items-center justify-center transition-colors"
-                    />
-                    <p className="flex-1 text-sm text-gray-300 font-mono truncate">
-                      {task.content}
-                    </p>
-                    <button
-                      onClick={() => deleteTask(task.id)}
-                      className="text-gray-700 hover:text-red-400 text-xs opacity-0 group-hover:opacity-100 transition-all"
-                    >
-                      {"✕"}
-                    </button>
-                  </motion.div>
-                ))}
-                {completedTasks.map((task) => (
-                  <motion.div
-                    key={task.id}
-                    layout
-                    initial={{ opacity: 0 }}
-                    animate={{ opacity: 1 }}
-                    exit={{ opacity: 0 }}
-                    className="flex items-center gap-2.5 group"
-                  >
-                    <button
-                      onClick={() => toggleTask(task.id, false)}
-                      className="flex-shrink-0 w-[18px] h-[18px] rounded border-2 border-emerald-600 bg-emerald-600 flex items-center justify-center"
-                    >
-                      <span className="text-white text-[10px]">{"✓"}</span>
-                    </button>
-                    <p className="flex-1 text-sm text-gray-600 font-mono truncate line-through">
-                      {task.content}
-                    </p>
-                    <button
-                      onClick={() => deleteTask(task.id)}
-                      className="text-gray-700 hover:text-red-400 text-xs opacity-0 group-hover:opacity-100 transition-all"
-                    >
-                      {"✕"}
-                    </button>
-                  </motion.div>
-                ))}
-              </AnimatePresence>
-            </div>
+          <FriendsOnlineSection
+            onlineFriends={onlineFriends}
+            onOpenFriends={onOpenFriends}
+            onJoinSession={onJoinSession}
+            onInvite={onInvite}
+          />
 
-            {tasks.length === 0 && (
-              <p className="text-gray-600 text-xs font-mono text-center py-3">
-                Add goals for your focus session
-              </p>
-            )}
-
-            {/* Add task input */}
-            <div className="flex gap-2 mt-3 pt-3 border-t border-gray-700">
-              <input
-                className="flex-1 bg-transparent text-sm text-gray-300 placeholder-gray-600 focus:outline-none font-mono"
-                placeholder="Add a goal..."
-                value={newTask}
-                onChange={(e) => setNewTask(e.target.value)}
-                onKeyDown={(e) => e.key === "Enter" && addTask()}
-                maxLength={120}
-              />
-              <button
-                onClick={addTask}
-                disabled={!newTask.trim()}
-                className="text-emerald-400 font-bold text-lg disabled:opacity-30 transition-opacity"
-              >
-                +
-              </button>
-            </div>
-          </div>
-
-          {/* Friends Online */}
-          {onlineFriends.length > 0 && (
-            <div className="bg-gray-800/50 rounded-2xl border border-gray-700 p-4">
-              <div className="flex items-center justify-between mb-3">
-                <h2 className="text-sm font-bold font-mono text-gray-400 uppercase tracking-wider">
-                  Friends Online
-                </h2>
-                <button
-                  onClick={onOpenFriends}
-                  className="text-[10px] font-mono text-gray-600 hover:text-emerald-400 transition-colors"
-                >
-                  See all
-                </button>
-              </div>
-              <div className="space-y-1">
-                {onlineFriends.slice(0, 5).map((f) => {
-                  const inSession = !!f.current_session_id;
-                  const worldInfo = f.current_world_id
-                    ? WORLD_LABEL[f.current_world_id]
-                    : null;
-                  const name = f.display_name ?? f.username;
-                  return (
-                    <div
-                      key={f.id}
-                      className="flex items-center gap-2.5 py-2 px-2.5 rounded-xl hover:bg-gray-700/50 transition-colors group"
-                    >
-                      <div
-                        className={`w-2 h-2 rounded-full flex-shrink-0 ${inSession ? "bg-emerald-400 animate-pulse" : "bg-yellow-400"}`}
-                      />
-                      <div className="flex-1 min-w-0">
-                        <p className="text-sm font-bold text-white truncate">
-                          {name}
-                        </p>
-                        {inSession && worldInfo ? (
-                          <p className="text-[10px] text-emerald-400 font-mono truncate">
-                            {worldInfo.emoji} In {worldInfo.label}
-                          </p>
-                        ) : (
-                          <p className="text-[10px] text-gray-500 font-mono">
-                            Online
-                          </p>
-                        )}
-                      </div>
-                      {inSession && f.current_session_id ? (
-                        <button
-                          onClick={() => onJoinSession(f.current_session_id!)}
-                          className="text-[10px] bg-emerald-500/20 hover:bg-emerald-500/40 text-emerald-300 font-mono font-bold px-2 py-1 rounded-lg transition-colors"
-                        >
-                          Join
-                        </button>
-                      ) : (
-                        <button
-                          onClick={() => onInvite(f.id)}
-                          className="text-[10px] bg-gray-700/50 hover:bg-gray-700 text-gray-400 font-mono font-bold px-2 py-1 rounded-lg transition-colors opacity-0 group-hover:opacity-100"
-                        >
-                          Invite
-                        </button>
-                      )}
-                    </div>
-                  );
-                })}
-                {onlineFriends.length > 5 && (
-                  <button
-                    onClick={onOpenFriends}
-                    className="w-full text-center text-[10px] font-mono text-gray-500 hover:text-emerald-400 py-1.5 transition-colors"
-                  >
-                    +{onlineFriends.length - 5} more
-                  </button>
-                )}
-              </div>
-            </div>
-          )}
-
-          {/* World Picker */}
           <div>
             <h2 className="text-sm font-bold font-mono text-gray-400 uppercase tracking-wider mb-3">
               Choose World
@@ -543,7 +277,6 @@ export default function HomeDashboard({
             </div>
           </div>
 
-          {/* Focus Button — directly below world picker */}
           <div className="space-y-2 pt-2">
             {activeSessionId && (
               <button
